requirements.txt
Django==4.2.7
django-environ==0.11.2
Pillow==10.1.0
qrcode[pil]==7.4.2
numpy==2.2.4
psutil==7.0.0
groq==0.20.0
python-dotenv==1.1.0
gunicorn==21.2.0
